Move slider and scrollbar implementations to separate cpp
