+++ CMakeLists.txt
@@ -17,7 +17,7 @@ add_library(reig_lib
         lib/reig/internal.h lib/reig/internal.cpp
         lib/reig/context.h lib/reig/context.cpp
         lib/reig/reference_widget_list.h lib/reig/reference_widget_list.cpp
-        )
+        lib/reig/reference_widget_slider.cpp)
 add_library(my::reig_lib ALIAS reig_lib)
 target_compile_features(reig_lib
         PUBLIC
